move rsonpath-benchmarks as a full subdirectory
We used submodules to separate rsonpath-benchmarks from rsonpath
but that is no longer useful. Since rsonpath-benchmarks needs to be updated
every time rsonpath is, and we want to run benchmarks every time we update rsonpath,
they are very much tied together. It remains separate from the main rsonpath
workspace, though, as its dependencies are completely separate.