chore: add info on how to get gradlew to work
After removing `gradle-wrapper.jar` from being tracked in the repo
you need to regenerate it manually when you clone the repo.
To do that, you need `gradle` actually installed and then to run `gradle wrapper`
in the correct directory. This was added to the README and to the benchmarks' `just init`.

Additionally updated the root Justfile to also perform that during `init`.