deps: bump nom to 8.0.0
There was a bunch of breaking changes in nom 8.0.0.
Nom's maintainers claim it reduces code generated by parsers,
which is always worth it.

It actually seems to improve performance by roughly 10%
on compiling small queries. On larger queries the cost of compilation
outweighs the cost of parsing so it's not noticable.