fix: fix panic in specific cases of reclassification at end of file (#788)
* fix: fix panic in specific cases of reclassification at end of file

A particular combination of reclassification after tail-skipping
at the very end of the file could cause a panic if the file-ending
closing occurred directly after the skipped-to character.

The root cause was that the depth-skipping code returned the byte
index one *after* the skipped-to character, which would then cause
the structural classifier to consider that position in the mask
to be irrelevant and zero it out during reclassification.

This commit adds a test case that reproduces the issue and then
fixes the off-by-one error.